Beef up cat tests
