+++ py-polars/tests/unit/operations/namespaces/test_categorical.py
@@ -1,36 +1,14 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
-
 import pytest
 
 import polars as pl
 from polars.testing import assert_frame_equal, assert_series_equal
 
-if TYPE_CHECKING:
-    from collections.abc import Generator
-    from typing import Any
-
-    FixtureRequest = Any
-
-
-@pytest.fixture(params=[True, False])
-def test_global_and_local(
-    request: FixtureRequest,
-) -> Generator[Any, Any, Any]:
-    """Setup fixture which runs each test with and without global string cache."""
-    use_global = request.param
-    if use_global:
-        with pl.StringCache():
-            # Pre-fill some global items to ensure physical repr isn't 0..n.
-            pl.Series(["a", "b", "c"], dtype=pl.Categorical)
-            yield
-    else:
-        yield
-
 
 # @TODO: Bug, see https://github.com/pola-rs/polars/issues/20440
 @pytest.mark.may_fail_auto_streaming
+@pytest.mark.usefixtures("test_global_and_local")
 def test_categorical_lexical_sort() -> None:
     df = pl.DataFrame(
         {"cats": ["z", "z", "k", "a", "b"], "vals": [3, 1, 2, 2, 3]}
@@ -66,50 +44,51 @@ def test_categorical_lexical_sort() -> None:
     ]
 
 
+@pytest.mark.usefixtures("test_global_and_local")
 def test_categorical_lexical_ordering_after_concat() -> None:
-    with pl.StringCache():
-        ldf1 = (
-            pl.DataFrame([pl.Series("key1", [8, 5]), pl.Series("key2", ["fox", "baz"])])
-            .lazy()
-            .with_columns(pl.col("key2").cast(pl.Categorical("lexical")))
+    ldf1 = (
+        pl.DataFrame([pl.Series("key1", [8, 5]), pl.Series("key2", ["fox", "baz"])])
+        .lazy()
+        .with_columns(pl.col("key2").cast(pl.Categorical("lexical")))
+    )
+    ldf2 = (
+        pl.DataFrame(
+            [pl.Series("key1", [6, 8, 6]), pl.Series("key2", ["fox", "foo", "bar"])]
         )
-        ldf2 = (
-            pl.DataFrame(
-                [pl.Series("key1", [6, 8, 6]), pl.Series("key2", ["fox", "foo", "bar"])]
-            )
-            .lazy()
-            .with_columns(pl.col("key2").cast(pl.Categorical("lexical")))
-        )
-        df = pl.concat([ldf1, ldf2]).select(pl.col("key2")).collect()
+        .lazy()
+        .with_columns(pl.col("key2").cast(pl.Categorical("lexical")))
+    )
+    df = pl.concat([ldf1, ldf2]).select(pl.col("key2")).collect()
 
-        assert df.sort("key2").to_dict(as_series=False) == {
-            "key2": ["bar", "baz", "foo", "fox", "fox"]
-        }
+    assert df.sort("key2").to_dict(as_series=False) == {
+        "key2": ["bar", "baz", "foo", "fox", "fox"]
+    }
 
 
 @pytest.mark.may_fail_auto_streaming
+@pytest.mark.usefixtures("test_global_and_local")
 def test_sort_categoricals_6014_internal() -> None:
-    with pl.StringCache():
-        # create basic categorical
-        df = pl.DataFrame({"key": ["bbb", "aaa", "ccc"]}).with_columns(
-            pl.col("key").cast(pl.Categorical)
-        )
+    # create basic categorical
+    df = pl.DataFrame({"key": ["bbb", "aaa", "ccc"]}).with_columns(
+        pl.col("key").cast(pl.Categorical)
+    )
 
     out = df.sort("key")
     assert out.to_dict(as_series=False) == {"key": ["bbb", "aaa", "ccc"]}
 
 
+@pytest.mark.usefixtures("test_global_and_local")
 def test_sort_categoricals_6014_lexical() -> None:
-    with pl.StringCache():
-        # create lexically-ordered categorical
-        df = pl.DataFrame({"key": ["bbb", "aaa", "ccc"]}).with_columns(
-            pl.col("key").cast(pl.Categorical("lexical"))
-        )
+    # create lexically-ordered categorical
+    df = pl.DataFrame({"key": ["bbb", "aaa", "ccc"]}).with_columns(
+        pl.col("key").cast(pl.Categorical("lexical"))
+    )
 
     out = df.sort("key")
     assert out.to_dict(as_series=False) == {"key": ["aaa", "bbb", "ccc"]}
 
 
+@pytest.mark.usefixtures("test_global_and_local")
 def test_categorical_get_categories() -> None:
     assert pl.Series(
         "cats", ["foo", "bar", "foo", "foo", "ham"], dtype=pl.Categorical
@@ -166,6 +145,7 @@ def test_cat_is_local() -> None:
     assert not s2.cat.is_local()
 
 
+@pytest.mark.usefixtures("test_global_and_local")
 def test_cat_uses_lexical_ordering() -> None:
     s = pl.Series(["a", "b", None, "b"]).cast(pl.Categorical)
     assert s.cat.uses_lexical_ordering() is False
